+++ src/prim/prim_mt.h
@@ -34,4 +34,4 @@
 
 #define MAX_THREADS 8
 
-int *prim_mt_mst(float *graph, int graph_size, int graph_root, int num_threads);
+int *prim_mt_mst(unsigned short int *graph, int graph_size, int graph_root, int num_threads);
